feat : formatWeek함수 테스트
