+++ packages/assignment-7/src/App.tsx
@@ -40,7 +40,7 @@ import {
   DeleteIcon,
   EditIcon,
 } from "@chakra-ui/icons";
-import { getDaysInMonth, getWeekDates } from "./utils/date-utils";
+import { formatWeek, getDaysInMonth, getWeekDates } from "./utils/date-utils";
 
 type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
 
@@ -430,13 +430,6 @@ function App() {
       return true;
     });
   })();
-
-  const formatWeek = (date: Date): string => {
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const weekNumber = Math.ceil(date.getDate() / 7);
-    return `${year}년 ${month}월 ${weekNumber}주`;
-  };
 
   const formatMonth = (date: Date): string => {
     const year = date.getFullYear();
